feat: enhance shield material system with registry and built-in materials
