+++ src/config/renderer.ts
@@ -30,11 +30,31 @@ export interface ShieldVisualSettings {
 
 // Tunable per-hull shield visuals; values are conservative defaults.
 export const SHIELD_VISUALS: Record<ShipHull, ShieldVisualSettings> = {
-  fighter:   { margin: 1.01, hexScale: 48, edgeWidth: 0.10, maxAlpha: 0.5 },
-  corvette:  { margin: 1.01, hexScale: 48, edgeWidth: 0.10, maxAlpha: 0.5 },
-  frigate:   { margin: 1.01, hexScale: 48, edgeWidth: 0.10, maxAlpha: 0.5 },
-  destroyer: { margin: 1.01, hexScale: 48, edgeWidth: 0.10, maxAlpha: 0.5 },
-  carrier:   { margin: 1.01, hexScale: 48, edgeWidth: 0.10, maxAlpha: 0.5 },
+  fighter:   { margin: 1.01, 
+               hexScale: 48, 
+               edgeWidth: 0.10, 
+               maxAlpha: 0.5,
+               materialKind: 'hex' },
+  corvette:  { margin: 1.01, 
+               hexScale: 48, 
+               edgeWidth: 0.10, 
+               maxAlpha: 0.5, 
+               materialKind: 'hex' },
+  frigate:   { margin: 1.01, 
+               hexScale: 48, 
+               edgeWidth: 0.10, 
+               maxAlpha: 0.5,
+               materialKind: 'hex' },
+  destroyer: { margin: 1.01, 
+               hexScale: 48, 
+               edgeWidth: 0.10, 
+               maxAlpha: 0.5,
+               materialKind: 'hex' },
+  carrier:   { margin: 1.01, 
+               hexScale: 48, 
+               edgeWidth: 0.10, 
+               maxAlpha: 0.5,
+               materialKind: 'hex' },
 };
 
 const DEFAULTS: Required<ShieldVisualSettings> = {
@@ -46,10 +66,10 @@ const DEFAULTS: Required<ShieldVisualSettings> = {
   transmission: {
     thickness: 0.6,
     chromaticAberration: 0.02,
-    anisotropicBlur: 0.0,
-    distortion: 0.0,
+    anisotropicBlur: 0.1,
+    distortion: 0.1,
     distortionScale: 0.4,
-    temporalDistortion: 0.0,
+    temporalDistortion: 0.1,
     attenuationDistance: 0.6,
     roughness: 0.1,
     clearcoat: 0.0,
